EDFS/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
config-parser==0.0.1
configparser==5.3.0
dnspython==2.2.1
idna==3.4
mysql-client==0.0.1
pymongo==4.3.3
pypi-xmlrpc==2020.12.3
PyYAML==6.0
requests==2.28.1
urllib3==1.26.13
